rename remaining test classes
