+++ integration-tests/src/test/kotlin/com/testapp/tests/payments/flows/paynow/PayNowSofortTest.kt
@@ -1,4 +1,4 @@
-package com.testapp.tests.payments.flows.sliceit
+package com.testapp.tests.payments.flows.paynow
 
 import com.testapp.base.BaseAppiumTest
 import com.testapp.base.PaymentCategory
@@ -8,34 +8,35 @@ import com.testapp.network.KlarnaApi
 import com.testapp.utils.*
 import io.appium.java_client.MobileElement
 import io.appium.java_client.android.AndroidDriver
+import io.appium.java_client.ios.IOSElement
 import org.junit.Assert
 import org.junit.Test
 import org.openqa.selenium.By
 import org.openqa.selenium.support.ui.ExpectedConditions
 
-internal class TestSliceItUK : BaseAppiumTest() {
+internal class PayNowSofortTest : BaseAppiumTest() {
 
     @Test
-    fun `test payment slice it UK successful flow`() {
-        testSliceItUK(true)
+    fun `test payment pay now DE Sofort successful flow`() {
+        testPayNowSofort(true)
     }
 
     @Test
-    fun `test payment slice it UK failure flow`() {
-        testSliceItUK(false)
+    fun `test payment pay now DE Sofort failure flow`() {
+        testPayNowSofort(false)
     }
 
-    private fun testSliceItUK(success: Boolean) {
-        val session = KlarnaApi.getSessionInfo(SessionHelper.getRequestUK())?.session
+    private fun testPayNowSofort(success: Boolean) {
+        val session = KlarnaApi.getSessionInfo(SessionHelper.getRequestDE())?.session
         if (session?.client_token == null || !session.payment_method_categories.map { it.identifier }
-                .contains(PaymentCategory.SLICE_IT.value)) {
+                .contains(PaymentCategory.PAY_NOW.value)) {
             return
         }
         val token = session.client_token
-        initLoadSDK(token, PaymentCategory.SLICE_IT.value)
+        initLoadSDK(token, PaymentCategory.PAY_NOW.value)
+
         if (android()) {
             DriverUtils.switchContextToWebView(driver)
-
             val mainWindow = WebViewTestHelper.findWindowFor(driver, By.id("klarna-some-hardcoded-instance-id-main"))
             mainWindow?.let {
                 driver.switchTo().window(it)
@@ -43,78 +44,72 @@ internal class TestSliceItUK : BaseAppiumTest() {
             DriverUtils.getWaiter(driver)
                 .until(ExpectedConditions.frameToBeAvailableAndSwitchToIt("klarna-some-hardcoded-instance-id-main"))
 
-            DriverUtils.waitForPresence(driver, By.id("scheme-payment-selector"))
-//
-//        PaymentFlowsTestHelper.fillCardInfo(driver)
+            DriverUtils.getWaiter(driver)
+                .until(ExpectedConditions.presenceOfElementLocated(By.id("installments-card|-1"))).click()
 
-            DriverUtils.switchContextToNative(driver)
+            DriverUtils.getWaiter(driver)
+                .until(ExpectedConditions.frameToBeAvailableAndSwitchToIt(By.xpath("//*[@id=\"pay-now-card\"]//iframe")))
+
         } else {
             DriverUtils.getWaiter(driver)
                 .until(ExpectedConditions.presenceOfElementLocated(By.xpath("//XCUIElementTypeOther[@name='Payment View']")))
+            val card: IOSElement = DriverUtils.getWaiter(driver)
+                .until(ExpectedConditions.presenceOfElementLocated(By.xpath("//XCUIElementTypeStaticText[@name='Card']"))) as IOSElement
+            card.tapElementCenter(driver)
+            DriverUtils.wait(driver, 5)
         }
+
+        PaymentFlowsTestHelper.fillCardInfo(driver)
+
+        DriverUtils.switchContextToNative(driver)
 
         driver.hideKeyboardCompat()
 
         PaymentFlowsTestHelper.dismissConsole(driver)
 
         try {
-            driver.findElement(ByRnId(driver, "authorizeButton_${PaymentCategory.SLICE_IT.value}")).click()
+            driver.findElement(ByRnId(driver, "authorizeButton_${PaymentCategory.PAY_NOW.value}")).click()
         } catch (t: Throwable) {
             (driver as? AndroidDriver<*>)?.let { driver ->
-                driver.findElementByAndroidUIAutomator("new UiScrollable(new UiSelector().scrollable(true).instance(0)).scrollIntoView(new UiSelector().description(\"authorizeButton_${PaymentCategory.SLICE_IT.value}\"))")
+                driver.findElementByAndroidUIAutomator("new UiScrollable(new UiSelector().scrollable(true).instance(0)).scrollIntoView(new UiSelector().description(\"authorizeButton_${PaymentCategory.PAY_NOW.value}\"))")
             }
             DriverUtils.getWaiter(driver).until(
                 ExpectedConditions.presenceOfElementLocated(
                     ByRnId(
                         driver,
-                        "authorizeButton_${PaymentCategory.SLICE_IT.value}"
+                        "authorizeButton_${PaymentCategory.PAY_NOW.value}"
                     )
                 )
             ).click()
         }
 
-        val billing = BillingAddressTestHelper.getBillingInfoUK()
+        val billing = BillingAddressTestHelper.getBillingInfoDE()
         if (!success) {
             BillingAddressTestHelper.setEmailFlag(billing, BillingAddressTestHelper.EMAIL_FLAG_REJECTED)
         }
         PaymentFlowsTestHelper.fillBillingAddress(driver, billing)
 
-        PaymentFlowsTestHelper.fillSmsCode(driver)
-
-        PaymentFlowsTestHelper.fillInfo(driver, billing, billing.identifiers.birthday2, billing.options.birthday)
-
-        if (android()) {
-            PaymentFlowsTestHelper.submitAndConfirm(
-                driver,
-                By.xpath("//button[contains(@id,'purchase-approval-form-continue-button')]"),
-                By.xpath("//div[contains(@id,'purchase-approval__footer-button-wrapper')]")
-            )
-        } else {
-            driver.hideKeyboardCompat()
-            PaymentFlowsTestHelper.submitAndConfirm(
-                driver,
-                By.xpath("//XCUIElementTypeButton[contains(@name,'Confirm')]"),
-                null
-            )
-        }
-
         if (!success) {
             if (android()) {
-                DriverUtils.waitForPresence(
-                    driver,
-                    By.xpath("//h1[contains(text(),'Your application was declined')]")
-                )
+                val refusedTextBy = By.xpath("//*[@id=\"message-component-root\"]")
+                val refusedText =
+                    DriverUtils.getWaiter(driver).until(ExpectedConditions.presenceOfElementLocated(refusedTextBy))
+                with(refusedText.text.toLowerCase()) {
+                    assert(this.contains("sorry") || this.contains("unfortunately"))
+                }
             } else {
                 val changePaymentBy = By.xpath("//XCUIElementTypeButton[contains(@name, 'Change payment method')]")
                 val changePayment =
                     DriverUtils.getWaiter(driver).until(ExpectedConditions.presenceOfElementLocated(changePaymentBy))
                 (changePayment as MobileElement).tapElementCenter(driver)
-                val response = PaymentFlowsTestHelper.readStateMessage(driver, PaymentCategory.SLICE_IT)
+                val response = PaymentFlowsTestHelper.readStateMessage(driver, PaymentCategory.PAY_NOW)
+
                 PaymentFlowsTestHelper.checkAuthorizeResponse(response, false)
             }
+
         } else {
             DriverUtils.switchContextToNative(driver)
-            val response = PaymentFlowsTestHelper.readStateMessage(driver, PaymentCategory.SLICE_IT)
+            val response = PaymentFlowsTestHelper.readStateMessage(driver, PaymentCategory.PAY_NOW)
 
             PaymentFlowsTestHelper.checkAuthorizeResponse(response, true)
         }
